moved RowOfOptions into its own widget
